Change the NETCONF TLS demo to a NETCONF RESTCONF demo
