+++ tls-netconf/app/cmd-set-dhcp-defaultLeaseTime-1h.xml
@@ -16,6 +16,11 @@
 ]]>]]>
 <?xml version="1.0" encoding="UTF-8"?>
 <rpc xmlns="urn:ietf:params:xml:ns:netconf:base:1.0" message-id="2">
+  <discard-changes/>
+</rpc>
+]]>]]>
+<?xml version="1.0" encoding="UTF-8"?>
+<rpc xmlns="urn:ietf:params:xml:ns:netconf:base:1.0" message-id="3">
   <edit-config>
     <target>
       <candidate/>
@@ -32,12 +37,12 @@
 </rpc>
 ]]>]]>
 <?xml version="1.0" encoding="UTF-8"?>
-<rpc xmlns="urn:ietf:params:xml:ns:netconf:base:1.0" message-id="3">
+<rpc xmlns="urn:ietf:params:xml:ns:netconf:base:1.0" message-id="4">
   <commit/>
 </rpc>
 ]]>]]>
 <?xml version="1.0" encoding="UTF-8"?>
-<rpc xmlns="urn:ietf:params:xml:ns:netconf:base:1.0" message-id="4">
+<rpc xmlns="urn:ietf:params:xml:ns:netconf:base:1.0" message-id="5">
   <unlock>
     <target>
       <candidate/>
@@ -46,7 +51,7 @@
 </rpc>
 ]]>]]>
 <?xml version="1.0" encoding="UTF-8"?>
-<rpc xmlns="urn:ietf:params:xml:ns:netconf:base:1.0" message-id="5">
+<rpc xmlns="urn:ietf:params:xml:ns:netconf:base:1.0" message-id="6">
   <close-session/>
 </rpc>
 ]]>]]>
