+++ tls-netconf/app/confd.conf
@@ -1,0 +1,227 @@
+<!-- -*- nxml -*- -->
+<!-- This configuration is good for the examples, but are in many ways
+     atypical for a production system. It also does not contain all
+     possible configuration options.
+
+     Better starting points for a production confd.conf configuration
+     file would be confd.conf.example. For even more information, see
+     the confd.conf man page.
+
+     E.g. references to current directory are not good practice in a
+     production system, but makes it easier to get started with
+     this example. There are many references to the current directory
+     in this example configuration.
+-->
+
+<confdConfig xmlns="http://tail-f.com/ns/confd_cfg/1.0">
+  <!-- The loadPath is searched for .fxs files, javascript files, etc.
+       NOTE: if you change the loadPath, the daemon must be restarted,
+       or the "In-service Data Model Upgrade" procedure described in
+       the User Guide can be used - 'confd - -reload' is not enough.
+  -->
+  <loadPath>
+    <dir>.</dir>
+  </loadPath>
+
+  <stateDir>.</stateDir>
+
+  <enableAttributes>true</enableAttributes>
+
+  <cdb>
+    <enabled>true</enabled>
+    <dbDir>./confd-cdb</dbDir>
+    <operational>
+      <enabled>true</enabled>
+    </operational>
+  </cdb>
+
+  <rollback>
+    <enabled>true</enabled>
+    <directory>./confd-cdb</directory>
+  </rollback>
+
+  <!-- These keys are used to encrypt values adhering to the types
+       tailf:des3-cbc-encrypted-string and tailf:aes-cfb-128-encrypted-string
+       as defined in the tailf-common YANG module. These types are
+       described in confd_types(3).
+  -->
+  <encryptedStrings>
+    <DES3CBC>
+      <key1>0123456789abcdef</key1>
+      <key2>0123456789abcdef</key2>
+      <key3>0123456789abcdef</key3>
+      <initVector>0123456789abcdef</initVector>
+    </DES3CBC>
+
+    <AESCFB128>
+      <key>0123456789abcdef0123456789abcdef</key>
+      <initVector>0123456789abcdef0123456789abcdef</initVector>
+    </AESCFB128>
+  </encryptedStrings>
+
+  <logs>
+    <!-- Shared settings for how to log to syslog.
+         Each log can be configured to log to file and/or syslog.  If a
+         log is configured to log to syslog, the settings below are used.
+    -->
+    <syslogConfig>
+      <!-- facility can be 'daemon', 'local0' ... 'local7' or an integer -->
+      <facility>daemon</facility>
+      <!-- if udp is not enabled, messages will be sent to local syslog -->
+      <udp>
+        <enabled>false</enabled>
+        <host>syslogsrv.example.com</host>
+        <port>514</port>
+      </udp>
+    </syslogConfig>
+
+    <!-- 'confdlog' is a normal daemon log.  Check this log for
+         startup problems of confd itself.
+         By default, it logs directly to a local file, but it can be
+         configured to send to a local or remote syslog as well.
+    -->
+    <confdLog>
+      <enabled>true</enabled>
+      <file>
+        <enabled>true</enabled>
+        <name>./confd.log</name>
+      </file>
+      <syslog>
+        <enabled>true</enabled>
+      </syslog>
+    </confdLog>
+
+    <!-- The developer logs are supposed to be used as debug logs
+         for troubleshooting user-written javascript and c code.  Enable
+         and check these logs for problems with validation code etc.
+    -->
+    <developerLog>
+      <enabled>true</enabled>
+      <file>
+        <enabled>true</enabled>
+        <name>./devel.log</name>
+      </file>
+      <syslog>
+        <enabled>false</enabled>
+      </syslog>
+    </developerLog>
+    <developerLogLevel>trace</developerLogLevel>
+
+    <auditLog>
+      <enabled>true</enabled>
+      <file>
+        <enabled>true</enabled>
+        <name>./audit.log</name>
+      </file>
+      <syslog>
+        <enabled>true</enabled>
+      </syslog>
+    </auditLog>
+
+    <!-- The netconf log can be used to troubleshoot NETCONF operations,
+         such as checking why e.g. a filter operation didn't return the
+         data requested.
+    -->
+    <netconfLog>
+      <enabled>false</enabled>
+      <file>
+        <enabled>true</enabled>
+        <name>./netconf.log</name>
+      </file>
+      <syslog>
+        <enabled>false</enabled>
+      </syslog>
+    </netconfLog>
+
+    <errorLog>
+      <enabled>true</enabled>
+      <filename>./error.log</filename>
+      <debug>
+        <enabled>true</enabled>
+      </debug>
+    </errorLog>
+
+    <netconfTraceLog>
+      <enabled>true</enabled>
+      <filename>./netconf.trace</filename>
+      <format>pretty</format>
+    </netconfTraceLog>
+
+  </logs>
+
+  <!-- Defines which datastores confd will handle. -->
+  <datastores>
+    <!-- 'startup' means that the system keeps separate running and
+         startup configuration databases.  When the system reboots for
+         whatever reason, the running config database is lost, and the
+         startup is read.
+         Enable this only if your system uses a separate startup and
+         running database.
+    -->
+    <startup>
+      <enabled>false</enabled>
+    </startup>
+
+    <!-- The 'candidate' is a shared, named alternative configuration
+         database which can be modified without impacting the running
+         configuration.  Changes in the candidate can be commit to running,
+         or discarded.
+         Enable this if you want your users to use this feature from
+         NETCONF, CLI or WebGUI, or other agents.
+    -->
+    <candidate>
+      <enabled>true</enabled>
+      <!-- By default, confd implements the candidate configuration
+           without impacting the application.  But if your system
+           already implements the candidate itself, set 'implementation' to
+           'external'.
+      -->
+      <!--implementation>external</implementation-->
+      <implementation>confd</implementation>
+      <storage>auto</storage>
+      <filename>./confd_candidate.db</filename>
+    </candidate>
+    <running>
+      <access>writable-through-candidate</access>
+    </running>
+  </datastores>
+
+  <aaa>
+    <sshServerKeyDir>./ssh-keydir</sshServerKeyDir>
+  </aaa>
+
+  <netconf>
+    <enabled>true</enabled>
+    <transport>
+      <ssh>
+        <enabled>false</enabled>
+        <ip>0.0.0.0</ip>
+        <port>2022</port>
+      </ssh>
+      <tcp>
+        <enabled>true</enabled>
+        <ip>127.0.0.1</ip>
+        <port>2023</port>
+      </tcp>
+    </transport>
+
+    <capabilities>
+      <candidate>
+        <enabled>true</enabled>
+      </candidate>
+
+      <confirmed-commit>
+        <enabled>true</enabled>
+      </confirmed-commit>
+
+      <writable-running>
+        <enabled>false</enabled>
+      </writable-running>
+
+      <rollback-on-error>
+        <enabled>true</enabled>
+      </rollback-on-error>
+
+    </capabilities>
+  </netconf>
+</confdConfig>
