+++ tls-restconf/app/confd.conf
@@ -3,10 +3,10 @@
      atypical for a production system. It also does not contain all
      possible configuration options.
 
-     Better starting points for a production confd.conf configuration 
-     file would be confd.conf.example. For even more information, see 
+     Better starting points for a production confd.conf configuration
+     file would be confd.conf.example. For even more information, see
      the confd.conf man page.
-     
+
      E.g. references to current directory are not good practice in a
      production system, but makes it easier to get started with
      this example. There are many references to the current directory
@@ -33,7 +33,7 @@
     <operational>
       <enabled>true</enabled>
     </operational>
-  </cdb> 
+  </cdb>
 
   <rollback>
     <enabled>true</enabled>
@@ -43,7 +43,7 @@
   <!-- These keys are used to encrypt values adhering to the types
        tailf:des3-cbc-encrypted-string and tailf:aes-cfb-128-encrypted-string
        as defined in the tailf-common YANG module. These types are
-       described in confd_types(3). 
+       described in confd_types(3).
   -->
   <encryptedStrings>
     <DES3CBC>
@@ -52,7 +52,7 @@
       <key3>0123456789abcdef</key3>
       <initVector>0123456789abcdef</initVector>
     </DES3CBC>
-    
+
     <AESCFB128>
       <key>0123456789abcdef0123456789abcdef</key>
       <initVector>0123456789abcdef0123456789abcdef</initVector>
@@ -81,7 +81,7 @@
          configured to send to a local or remote syslog as well.
     -->
     <confdLog>
-      <enabled>false</enabled>
+      <enabled>true</enabled>
       <file>
         <enabled>true</enabled>
         <name>./confd.log</name>
@@ -108,7 +108,7 @@
     <developerLogLevel>trace</developerLogLevel>
 
     <auditLog>
-      <enabled>false</enabled>
+      <enabled>true</enabled>
       <file>
         <enabled>true</enabled>
         <name>./audit.log</name>
@@ -133,13 +133,17 @@
       </syslog>
     </netconfLog>
 
-    <webuiBrowserLog>
-      <enabled>false</enabled>
-      <filename>./browser.log</filename>
-    </webuiBrowserLog>
+    <errorLog>
+      <enabled>true</enabled>
+      <filename>./error.log</filename>
+      <debug>
+        <enabled>true</enabled>
+      </debug>
+    </errorLog>
 
     <webuiAccessLog>
-      <enabled>false</enabled>
+      <enabled>true</enabled>
+      <trafficLog>true</trafficLog>
       <dir>./</dir>
     </webuiAccessLog>
 
@@ -183,7 +187,7 @@
       <storage>auto</storage>
       <filename>./confd_candidate.db</filename>
     </candidate>
-    <running>  
+    <running>
       <access>writable-through-candidate</access>
     </running>
   </datastores>
@@ -196,12 +200,12 @@
     <enabled>true</enabled>
     <transport>
       <ssh>
-        <enabled>false</enabled>
+        <enabled>true</enabled>
         <ip>0.0.0.0</ip>
         <port>2022</port>
       </ssh>
       <tcp>
-        <enabled>true</enabled>
+        <enabled>false</enabled>
         <ip>127.0.0.1</ip>
         <port>2023</port>
       </tcp>
@@ -226,23 +230,34 @@
 
     </capabilities>
   </netconf>
+
+  <restconf>
+    <enabled>true</enabled>
+    <requireModuleName>
+      <enabled>true</enabled>
+    </requireModuleName>
+  </restconf>
+
   <webui>
     <enabled>true</enabled>
-    
+
     <transport>
       <tcp>
-        <enabled>true</enabled>
+        <enabled>false</enabled>
         <ip>0.0.0.0</ip>
         <port>8008</port>
       </tcp>
-      
+
       <ssl>
         <enabled>true</enabled>
         <ip>0.0.0.0</ip>
         <port>8888</port>
+        <readFromDb>true</readFromDb>
+        <protocols>tlsv1.2</protocols>
+        <verify>3</verify>
       </ssl>
     </transport>
-    
+
     <cgi>
       <enabled>true</enabled>
       <php>
@@ -251,4 +266,3 @@
     </cgi>
   </webui>
 </confdConfig>
-
